menu item mapping for sales 2

+++ src/main/java/leoric/pizzacipollastorage/DTOs/MenuItemSale/MenuItemSaleCreateDto.java
@@ -9,6 +9,6 @@ import java.util.UUID;
 public class MenuItemSaleCreateDto {
     private UUID menuItemId;
     private DishSize dishSize;
-    private int quantitySold;
+    private Float quantitySold;
     private String cookName;
 }
