+++ src/main/java/leoric/pizzacipollastorage/DTOs/MenuItemSale/MenuItemSaleResponseDto.java
@@ -17,7 +17,7 @@ public class MenuItemSaleResponseDto {
     private UUID id;
     private String menuItem;
     private DishSize dishSize;
-    private int quantitySold;
+    private Float quantitySold;
     private String cookName;
     private LocalDateTime saleDate;
 }
